test: unify e2e test suite
