+++ packages/playground/server-components-worker/src/components/ReadServerState.client.jsx
@@ -1,0 +1,11 @@
+import {useServerState} from '@shopify/hydrogen/client';
+
+export default function ReadServerState() {
+  const {serverState} = useServerState();
+
+  return (
+    <div>
+      <p id="server-state">Pathname: {serverState.pathname}</p>
+    </div>
+  );
+}
