+++ packages/playground/server-components-worker/src/pages/test-server-state.server.jsx
@@ -1,0 +1,10 @@
+import ReadServerState from '../components/ReadServerState.client';
+
+export default function TestServerState() {
+  return (
+    <div>
+      <h1>Test Server State</h1>
+      <ReadServerState />
+    </div>
+  );
+}
