+++ packages/playground/server-components/.env.development
@@ -1,1 +1,2 @@
-PRIVATE_VARIABLE="42"
+PRIVATE_VARIABLE="42-private"
+PUBLIC_VARIABLE="42-public"
