+++ packages/playground/server-components/src/pages/index.server.jsx
@@ -1,4 +1,5 @@
 import {Link} from '@shopify/hydrogen';
+import ClientEnv from '../components/ClientEnv.client';
 
 export function api() {
   return new Response('some api response');
@@ -12,8 +13,13 @@ export default function Index() {
         About
       </Link>
 
-      <div className="secrets">
-        PRIVATE_VARIABLE:{Oxygen.env.PRIVATE_VARIABLE}
+      <div className="secrets-server">
+        <div>PUBLIC_VARIABLE:{import.meta.env.PUBLIC_VARIABLE || ''}|</div>
+        <div>PRIVATE_VARIABLE:{Oxygen.env.PRIVATE_VARIABLE || ''}|</div>
+      </div>
+
+      <div className="secrets-client">
+        <ClientEnv />
       </div>
     </>
   );
